% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-mixpower.R
\name{as_tibble.mp_power}
\alias{as_tibble.mp_power}
\title{Coerce mixpower results to a tibble}
\usage{
\method{as_tibble}{mp_power}(x, ...)
}
\arguments{
\item{x}{An \code{mp_power}, \code{mp_sensitivity}, or \code{mp_power_curve} object.}

\item{...}{Passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}} on the underlying data frame.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Requires the \pkg{tibble} package (installed with the tidyverse).
}
