% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model for a single simulated dataset}
\usage{
fit_model(data, formula)
}
\arguments{
\item{data}{A data.frame of simulated data.}

\item{formula}{A model formula.}
}
\value{
A fitted model object.
}
\description{
Fit a model for a single simulated dataset
}
