% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{run_parallel}
\alias{run_parallel}
\title{Placeholder for parallel execution}
\usage{
run_parallel(fun, ...)
}
\arguments{
\item{fun}{Function to run.}

\item{...}{Additional arguments to pass to fun.}
}
\value{
The result of fun.
}
\description{
Placeholder for parallel execution
}
