% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity analysis for simulated results}
\usage{
sensitivity(results, by)
}
\arguments{
\item{results}{A data.frame with columns effect and power.}

\item{by}{Column name to aggregate by.}
}
\value{
A data.frame of aggregated power.
}
\description{
Sensitivity analysis for simulated results
}
