% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-glmm-nb.R
\name{simulate_glmm_nb_data}
\alias{simulate_glmm_nb_data}
\title{Simulate count outcome data for a Negative Binomial GLMM}
\usage{
simulate_glmm_nb_data(
  scenario,
  predictor = "condition",
  subject = "subject",
  outcome = "y",
  item = NULL,
  theta = NULL
)
}
\arguments{
\item{scenario}{An \code{mp_scenario} object.}

\item{predictor}{Predictor column name.}

\item{subject}{Subject ID column name.}

\item{outcome}{Outcome column name.}

\item{item}{Optional item ID column name.}

\item{theta}{NB dispersion parameter (size). Larger = less over-dispersion.}
}
\value{
A data.frame with outcome and predictors.
}
\description{
Simulate count outcome data for a Negative Binomial GLMM
}
