% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_simulations}
\alias{summarize_simulations}
\title{Summarize simulation outputs}
\usage{
summarize_simulations(simulations)
}
\arguments{
\item{simulations}{A data.frame of simulations.}
}
\value{
A summary data.frame.
}
\description{
Summarize simulation outputs
}
