% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_effect.R
\name{test_effect}
\alias{test_effect}
\title{Extract a test statistic for a model term}
\usage{
test_effect(fit, term)
}
\arguments{
\item{fit}{A fitted model object.}

\item{term}{Term name to test.}
}
\value{
A data.frame with coefficient information.
}
\description{
Extract a test statistic for a model term
}
