% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-lme4.R
\name{mp_scenario_lme4}
\alias{mp_scenario_lme4}
\title{Convenience constructor for an lme4-based scenario}
\usage{
mp_scenario_lme4(
  formula,
  design,
  assumptions,
  predictor = "condition",
  subject = "subject",
  item = NULL,
  ...
)
}
\arguments{
\item{formula}{A mixed-model formula compatible with \code{lme4::lmer}.}

\item{design}{\code{mp_design()}.}

\item{assumptions}{\code{mp_assumptions()}.}

\item{predictor}{Name of fixed effect to test (must match fixed effect column name).}

\item{subject}{Subject grouping factor name.}

\item{item}{Optional item grouping factor name.}

\item{...}{Passed to \code{mp_backend_lme4()}.}
}
\value{
An \code{mp_scenario} with engine functions filled.
}
\description{
Convenience constructor for an lme4-based scenario
}
