% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-lme4-poisson.R
\name{mp_scenario_lme4_poisson}
\alias{mp_scenario_lme4_poisson}
\title{Create a fully specified MixPower scenario with the Poisson lme4 backend}
\usage{
mp_scenario_lme4_poisson(
  formula,
  design,
  assumptions,
  predictor = "condition",
  subject = "subject",
  outcome = "y",
  item = NULL,
  test_term = predictor,
  test_method = c("wald", "lrt"),
  null_formula = NULL
)
}
\arguments{
\item{formula}{Model formula.}

\item{design}{A \code{mp_design} object.}

\item{assumptions}{A \code{mp_assumptions} object.}

\item{predictor}{Predictor column name.}

\item{subject}{Subject ID column name.}

\item{outcome}{Outcome column name.}

\item{item}{Optional item ID column name.}

\item{test_term}{Optional explicit term to test. Defaults to \code{predictor}.}

\item{test_method}{Inference method (\code{"wald"} or \code{"lrt"}).}

\item{null_formula}{Optional null model formula for \code{"lrt"} tests.}
}
\value{
An object of class \code{mp_scenario}.
}
\description{
Create a fully specified MixPower scenario with the Poisson lme4 backend
}
