% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot.mp_sensitivity}
\alias{plot.mp_sensitivity}
\title{Plot a one-dimensional sensitivity curve}
\usage{
\method{plot}{mp_sensitivity}(x, y = c("estimate", "failure_rate"), ...)
}
\arguments{
\item{x}{An \code{mp_sensitivity} object.}

\item{y}{What to plot on the y-axis (\code{"estimate"} or \code{"failure_rate"}).}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns the plotted data.
}
\description{
Plot a one-dimensional sensitivity curve
}
