% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-glmm.R
\name{simulate_glmm_binomial_data}
\alias{simulate_glmm_binomial_data}
\title{Simulate binary outcome data for a GLMM with random intercepts}
\usage{
simulate_glmm_binomial_data(
  scenario,
  predictor = "condition",
  subject = "subject",
  outcome = "y",
  item = NULL
)
}
\arguments{
\item{scenario}{An \code{mp_scenario} object.}

\item{predictor}{Predictor column name.}

\item{subject}{Subject ID column name.}

\item{outcome}{Outcome column name.}

\item{item}{Optional item ID column name.}
}
\value{
A data.frame with outcome and predictors.
}
\description{
Simulate binary outcome data for a GLMM with random intercepts
}
